// src/app/models/top-picks-request-payload.ts
export class TopPicksListRequestPayload {
    AppId: number;
    CategoryId: string;
    CountryId: string;
    IsClub: number;
    IsFavorite: number;
    IsStock: number;
    KeyWord: string;
    PriceRangeId: string;
    PageNumber: number;
    PageSize: number;
    RegionId: string;
    SessionId: string;
    SizeId: string;
    StoreId: number;
    TypeId: string;
    UserId: number;
    VarietalId: string;
}